+++ libs/data-mapper-v2/src/components/canvas/ReactFlow.tsx
@@ -1,19 +1,8 @@
 import type { AppDispatch, RootState } from '../../core/state/Store';
-import { useEffect, useMemo, useRef, useCallback, useState, type MouseEvent } from 'react';
+import { useEffect, useRef, useCallback, useState, type MouseEvent } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import type {
-  Connection,
-  Node,
-  Edge,
-  ConnectionLineComponent,
-  NodeProps,
-  NodeTypes,
-  OnNodeDrag,
-  IsValidConnection,
-  EdgeChange,
-  XYPosition,
-} from '@xyflow/react';
-import { PanOnScrollMode, ReactFlow, addEdge, applyEdgeChanges, useEdges, useReactFlow } from '@xyflow/react';
+import type { Connection, Edge, ConnectionLineComponent, NodeTypes, OnNodeDrag, IsValidConnection, XYPosition } from '@xyflow/react';
+import { PanOnScrollMode, ReactFlow, addEdge, useEdges, useReactFlow } from '@xyflow/react';
 import { reactFlowStyle, useStyles } from './styles';
 import SchemaNode from '../common/reactflow/SchemaNode';
 import ConnectionLine from '../common/reactflow/edges/ConnectionLine';
@@ -31,161 +20,54 @@ import { FunctionNode } from '../common/reactflow/FunctionNode';
 import { useDrop } from 'react-dnd';
 import useResizeObserver from 'use-resize-observer';
 import type { Bounds } from '../../core';
-import { convertWholeDataMapToLayoutTree, isSourceNode, isTargetNode } from '../../utils/ReactFlow.Util';
-import { createEdgeId, createTemporaryEdgeId, splitEdgeId } from '../../utils/Edge.Utils';
+import { splitEdgeId } from '../../utils/Edge.Utils';
 import useAutoLayout from '../../ui/hooks/useAutoLayout';
-import cloneDeep from 'lodash/cloneDeep';
 import EdgePopOver from './EdgePopOver';
-import { getReactFlowNodeId } from '../../utils/Schema.Utils';
 import { getFunctionNode } from '../../utils/Function.Utils';
 import LoopEdge from '../common/reactflow/edges/LoopEdge';
 import { emptyCanvasRect } from '@microsoft/logic-apps-shared';
 import CanvasNode from '../common/reactflow/CanvasNode';
 import IntermediateConnectedEdge from '../common/reactflow/edges/IntermediateConnectedEdge';
+import useReactFlowStates from './useReactflowStates';
 interface DMReactFlowProps {
   setIsMapStateDirty?: (isMapStateDirty: boolean) => void;
 }
+
+const nodeTypes = {
+  schemaNode: SchemaNode,
+  functionNode: FunctionNode,
+  canvasNode: CanvasNode,
+} as NodeTypes;
+
+const edgeTypes = {
+  connectedEdge: ConnectedEdge,
+  loopEdge: LoopEdge,
+  intermediateConnectedEdge: IntermediateConnectedEdge,
+};
 
 export const ReactFlowWrapper = ({ setIsMapStateDirty }: DMReactFlowProps) => {
   const styles = useStyles();
   const reactFlowInstance = useReactFlow();
   const ref = useRef<HTMLDivElement>(null);
   const dispatch = useDispatch<AppDispatch>();
-  const edges = useEdges();
-  const {
-    sourceNodesMap,
-    targetNodesMap,
-    functionNodes,
-    flattenedSourceSchema,
-    flattenedTargetSchema,
-    dataMapConnections,
-    sourceStateConnections,
-    nodesForScroll,
-    temporaryEdgeMapping,
-  } = useSelector((state: RootState) => state.dataMap.present.curDataMapOperation);
+  const reactflowEdges = useEdges();
+  const { functionNodes } = useSelector((state: RootState) => state.dataMap.present.curDataMapOperation);
   const currentCanvasRect = useSelector(
     (state: RootState) => state.dataMap.present.curDataMapOperation.loadedMapMetadata?.canvasRect ?? emptyCanvasRect
   );
   const { width: currentWidth, height: currentHeight, x: currentX, y: currentY } = currentCanvasRect;
-  const [functionNodesForDragDrop, setFunctionNodesForDragDrop] = useState<Node[]>([]);
   const { width: newWidth = undefined, height: newHeight = undefined } = useResizeObserver<HTMLDivElement>({
     ref,
   });
   const [edgePopoverBounds, setEdgePopoverBounds] = useState<Bounds>();
 
-  const realEdges: Edge[] = useMemo(() => {
-    let edges: Edge[] = [];
-    if (Object.entries(dataMapConnections).length > 0) {
-      const layout = convertWholeDataMapToLayoutTree(flattenedSourceSchema, flattenedTargetSchema, functionNodes, dataMapConnections);
-      edges = layout.edges.map((edge) => {
-        const newEdge: Edge = {
-          id: createEdgeId(edge.sourceId, edge.targetId),
-          source: edge.sourceId,
-          target: edge.targetId,
-          type: edge.isRepeating ? 'loopEdge' : 'connectedEdge',
-          reconnectable: 'target',
-          zIndex: 150,
-          interactionWidth: 10,
-          data: { isRepeating: edge.isRepeating },
-          focusable: true,
-          deletable: true,
-          selectable: true,
-        };
-        return newEdge;
-      });
-    }
-
-    return edges;
-  }, [dataMapConnections, flattenedSourceSchema, flattenedTargetSchema, functionNodes]);
-
-  // Edges created when node is expanded/Collapsed
-  const temporaryEdgesMapForCollapsedNodes: Record<string, Edge> = useMemo(() => {
-    const newEdgesMap: Record<string, Edge> = {};
-    const sourceStateConnectionsEntries = Object.entries(sourceStateConnections);
-
-    if (sourceStateConnectionsEntries.length > 0) {
-      for (const entry of sourceStateConnectionsEntries) {
-        const sourceNodeId = getReactFlowNodeId(entry[0], true);
-        const targetIds = Object.keys(entry[1]);
-        for (const targetId of targetIds) {
-          const targetNodeId = getReactFlowNodeId(targetId, false);
-          const id = createTemporaryEdgeId(sourceNodeId, targetNodeId);
-          const edge: Edge = {
-            id: id,
-            source: sourceNodeId,
-            target: targetNodeId,
-            type: 'connectedEdge',
-            reconnectable: 'target',
-            focusable: true,
-            interactionWidth: 10,
-            deletable: true,
-            selectable: true,
-            zIndex: 150,
-            data: {
-              isIntermediate: true,
-            },
-          };
-          newEdgesMap[id] = edge;
-        }
-      }
-    }
-
-    return newEdgesMap;
-  }, [sourceStateConnections]);
-
-  // Edges created when node is expanded/Collapsed
-  const temporaryEdgesMapForScrolledNodes: Record<string, Edge> = useMemo(() => {
-    const newEdgesMap: Record<string, Edge> = {};
-    const connectionMapForSource = Object.entries(temporaryEdgeMapping);
-
-    for (const [id, sourceMap] of connectionMapForSource) {
-      for (const edgeId of Object.keys(sourceMap)) {
-        const splitNodeId = splitEdgeId(edgeId);
-        if (splitNodeId.length >= 2) {
-          /**
-           * if intermediate edge is connected to a target-schema node, then source for the edge should be the intermediate node
-           * if intermediate edge is connected to a source-schema node, then source for the edge should be the source schema node
-           * if intermediate edge is connected to a function node and the scrolled out node is source, i.e. this is on the input side, then source for the edge should be the intermediate node
-           * For all other cases, source-schema/function node should be the source for the edge
-           */
-          const [source, target] = isTargetNode(splitNodeId[0])
-            ? [splitNodeId[1], splitNodeId[0]]
-            : isSourceNode(splitNodeId[0])
-              ? [splitNodeId[0], splitNodeId[1]]
-              : isSourceNode(id)
-                ? [splitNodeId[1], splitNodeId[0]]
-                : [splitNodeId[0], splitNodeId[1]];
-
-          const edge: Edge = {
-            id: edgeId,
-            source: source,
-            target: target,
-            focusable: true,
-            deletable: true,
-            selectable: true,
-            type: 'intermediateConnectedEdge',
-            zIndex: 150,
-            interactionWidth: 10,
-            data: {
-              isIntermediate: true,
-              componentId: id,
-            },
-          };
-          newEdgesMap[edgeId] = edge;
-        }
-      }
-    }
-
-    return newEdgesMap;
-  }, [temporaryEdgeMapping]);
-
   useAutoLayout();
+  const { nodes, edges, setFunctionNodes } = useReactFlowStates({});
 
   useEffect(() => {
     const functionNodesForDragDrop = Object.entries(functionNodes).map(([key, functionData]) =>
       getFunctionNode(functionData, key, functionData.position)
     );
-    setFunctionNodesForDragDrop(functionNodesForDragDrop);
 
     if (ref?.current && newWidth !== undefined && newHeight !== undefined) {
       const { x: newX, y: newY, width: newWidth, height: newHeight } = ref.current.getBoundingClientRect();
@@ -222,59 +104,7 @@ export const ReactFlowWrapper = ({ setIsMapStateDirty }: DMReactFlowProps) => {
     }
   }, [functionNodes, newWidth, currentWidth, newHeight, currentHeight, currentX, currentY, dispatch, ref]);
 
-  useEffect(() => {
-    const edgeChanges: Record<string, EdgeChange> = {};
-    const allTemporaryConnections = {
-      ...temporaryEdgesMapForCollapsedNodes,
-      ...temporaryEdgesMapForScrolledNodes,
-    };
-
-    for (const [id, edge] of Object.entries(allTemporaryConnections)) {
-      edgeChanges[id] = {
-        type: 'add',
-        item: edge,
-      };
-    }
-
-    for (const edge of edges) {
-      if (edge.data?.isTemporary) {
-        const id = edge.id;
-        if (allTemporaryConnections[id]) {
-          delete edgeChanges[id];
-        } else {
-          edgeChanges[id] = {
-            type: 'remove',
-            id: id,
-          };
-        }
-      }
-    }
-
-    if (Object.entries(edgeChanges).length > 0) {
-      applyEdgeChanges(Object.values(edgeChanges), edges);
-    }
-  }, [edges, temporaryEdgesMapForCollapsedNodes, temporaryEdgesMapForScrolledNodes]);
-
   const isMapStateDirty = useSelector((state: RootState) => state.dataMap.present.isDirty);
-
-  const nodeTypes: Record<string, React.ComponentType<NodeProps>> = useMemo(
-    () =>
-      ({
-        schemaNode: SchemaNode,
-        functionNode: FunctionNode,
-        canvasNode: CanvasNode,
-      }) as NodeTypes,
-    []
-  );
-
-  const edgeTypes = useMemo(
-    () => ({
-      connectedEdge: ConnectedEdge,
-      loopEdge: LoopEdge,
-      intermediateConnectedEdge: IntermediateConnectedEdge,
-    }),
-    []
-  );
 
   const onEdgeConnect = useCallback(
     (connection: Connection) => {
@@ -289,7 +119,7 @@ export const ReactFlowWrapper = ({ setIsMapStateDirty }: DMReactFlowProps) => {
           zIndex: 150,
           interactionWidth: 10,
         },
-        edges
+        reactflowEdges
       );
 
       // danielle maybe get the input number from here?
@@ -301,14 +131,14 @@ export const ReactFlowWrapper = ({ setIsMapStateDirty }: DMReactFlowProps) => {
       dispatch(makeConnectionFromMap(connectionAction));
       dispatch(setSelectedItem(connection.target));
     },
-    [edges, dispatch]
+    [reactflowEdges, dispatch]
   );
 
   const isValidConnection: IsValidConnection = useCallback(
     (connection) => {
-      return !edges.find((edge) => edge.source === connection.source && edge.target === connection.target);
+      return !reactflowEdges.find((edge) => edge.source === connection.source && edge.target === connection.target);
     },
-    [edges]
+    [reactflowEdges]
   );
 
   // NOTE: Putting this useEffect here for vis next to onSave
@@ -344,12 +174,12 @@ export const ReactFlowWrapper = ({ setIsMapStateDirty }: DMReactFlowProps) => {
 
   const onFunctionNodeDrag: OnNodeDrag = useCallback(
     (_event, node, _nodes) => {
-      setFunctionNodesForDragDrop((functionNodesForDragDrop) => [
+      setFunctionNodes((functionNodesForDragDrop) => [
         ...functionNodesForDragDrop.filter((nodeFromState) => nodeFromState.id !== node.id),
         node,
       ]);
     },
-    [setFunctionNodesForDragDrop]
+    [setFunctionNodes]
   );
 
   const onFunctionNodeDragStop: OnNodeDrag = useCallback(
@@ -373,7 +203,7 @@ export const ReactFlowWrapper = ({ setIsMapStateDirty }: DMReactFlowProps) => {
         if (edge.data?.componentId && splitIds.length >= 2) {
           const directionId1 = edge.data?.componentId as string;
           const directionId2 = splitIds[0];
-          const directEdge = edges.find(
+          const directEdge = reactflowEdges.find(
             (edge) =>
               (edge.source === directionId1 && edge.target === directionId2) ||
               (edge.source === directionId2 && edge.target === directionId1)
@@ -394,17 +224,7 @@ export const ReactFlowWrapper = ({ setIsMapStateDirty }: DMReactFlowProps) => {
         });
       }
     },
-    [dispatch, setEdgePopoverBounds, edges]
-  );
-
-  const nodes = useMemo(
-    () => [
-      ...Object.values(sourceNodesMap),
-      ...Object.values(targetNodesMap),
-      ...functionNodesForDragDrop,
-      ...Object.values(nodesForScroll),
-    ],
-    [sourceNodesMap, targetNodesMap, functionNodesForDragDrop, nodesForScroll]
+    [dispatch, setEdgePopoverBounds, reactflowEdges]
   );
 
   return (
@@ -414,8 +234,8 @@ export const ReactFlowWrapper = ({ setIsMapStateDirty }: DMReactFlowProps) => {
           id="dm-react-flow"
           ref={drop}
           className="nopan nodrag"
-          nodes={cloneDeep(nodes)}
-          edges={[...realEdges, ...Object.values(temporaryEdgesMapForCollapsedNodes), ...Object.values(temporaryEdgesMapForScrolledNodes)]}
+          nodes={nodes}
+          edges={edges}
           nodeDragThreshold={0}
           onlyRenderVisibleElements={false}
           zoomOnScroll={false}
